+++ companion/public/manifest.json
@@ -4,8 +4,14 @@
   "version": "0.0.1",
   "description": "Companion for the Val Town CLI",
   "background": {
-    "service_worker": "worker/daemon.js"
+    "service_worker": "worker/worker.js"
   },
+  "content_scripts": [
+    {
+      "matches": ["<all_urls>"],
+      "js": ["content.js"]
+    }
+  ],
   "icons": {
     "16": "/assets/logo16.png",
     "32": "/assets/logo32.png",
@@ -13,5 +19,8 @@
     "128": "/assets/logo128.png"
   },
   "minimum_chrome_version": "116",
-  "permissions": ["tabs"]
+  "permissions": [
+    "tabs",
+    "contextMenus"
+  ]
 }
