+++ companion/src/content.ts
@@ -1,0 +1,7 @@
+window.onload = () => {
+  console.log("hi");
+};
+
+chrome.runtime.sendMessage({ action: "pageLoaded" }, (response) => {
+  console.log("Service worker response:", response);
+});
